rename nomic logging file
